simulator/requirements.txt
# -----------------------------------------------------------------------------
# 파일명 : simulator/requirements.txt
# 목적   : 시뮬레이터가 필요로 하는 파이썬 패키지 목록
# 비고   :
#   - PyYAML: profiles/routes 로딩
#   - confluent-kafka: Kafka 프로듀서(있으면 사용, 없으면 DummyProducer로 폴백)
#   - (선택) typing-extensions 등은 최신 파이썬에선 불필요
#   - 버전은 안정 버전 권장. 특정 조직 규정에 맞춰 핀(PIN) 가능.
# -----------------------------------------------------------------------------

PyYAML>=6.0.1,<7.0

# Kafka 드라이버 (옵션)
# 설치가 어려운 환경이라면 이 줄을 주석 처리해도 시뮬레이터는 DummyProducer로 동작합니다.
confluent-kafka>=2.5,<3.0
